+++ CHANGELOG.md
@@ -19,6 +19,11 @@ For older changes see the [archived Singularity change log](https://github.com/a
   fixes a problem seen when squashfuse (which is read-only) was used for
   the overlay layer.
 
+### New Features & Functionality
+
+- Added ability to set a custom config directory via the new
+  `APPTAINER_CONFIGDIR` environment variable.
+
 ## v1.2.0-rc.1 - \[2023-06-07\]
 
 ### Changed defaults / behaviours